MAINT split task and flow files into separate files
