+++ openml/runs/__init__.py
@@ -1,7 +1,7 @@
 from .run import OpenMLRun
-from .run import (run_task, get_run, list_runs, list_runs_by_flow,
-                  list_runs_by_tag, list_runs_by_task, list_runs_by_uploader,
-                  list_runs_by_filters)
+from .functions import (run_task, get_run, list_runs, list_runs_by_flow,
+                        list_runs_by_tag, list_runs_by_task,
+                        list_runs_by_uploader, list_runs_by_filters)
 
 __all__ = ['OpenMLRun', 'run_task', 'get_run', 'list_runs', 'list_runs_by_flow',
            'list_runs_by_tag', 'list_runs_by_task', 'list_runs_by_uploader',
