+++ tests/runs/test_run_functions.py
@@ -19,11 +19,11 @@ class TestRun(TestBase):
         clf = SGDClassifier(loss='hinge', random_state=1)
         self.assertRaisesRegexp(AttributeError,
                                 "probability estimates are not available for loss='hinge'",
-                                openml.runs.run._run_task_get_arffcontent,
+                                openml.runs.functions._run_task_get_arffcontent,
                                 clf, task, class_labels)
 
         clf = SGDClassifier(loss='log', random_state=1)
-        arff_datacontent = openml.runs.run._run_task_get_arffcontent(
+        arff_datacontent = openml.runs.functions._run_task_get_arffcontent(
             clf, task, class_labels)
         self.assertIsInstance(arff_datacontent, list)
         # 10 times 10 fold CV of 150 samples
